Group together id3 tag related code
